created first 10 test components w/ highlights
